Should be decimal

+++ StormCommerce.ExportFeed.Contracts/OnHand.cs
@@ -20,6 +20,6 @@ namespace StormCommerce.ExportFeed.Contracts
 		[DataMember(Order = 6)]
 		public DateTime? NextDeliveryDate { get; set; }
         [DataMember(Order = 7)]
-        public int? IncomingValue { get; set; }
+        public decimal? IncomingValue { get; set; }
     }
 }
